Move filter_results to filter/result
