+++ sdk/filter/result/filter_result.go
@@ -1,5 +1,4 @@
-package filterutils
-
+package result
 type FilterResult struct {
 	Block              bool
 	ResponseStatusCode int32
