+++ sdk/instrumentation/net/http/handler_test.go
@@ -9,7 +9,7 @@ import (
 
 	config "github.com/hypertrace/agent-config/gen/go/v1"
 	"github.com/hypertrace/goagent/sdk"
-	"github.com/hypertrace/goagent/sdk/filterutils"
+	"github.com/hypertrace/goagent/sdk/filter/result"
 	internalconfig "github.com/hypertrace/goagent/sdk/internal/config"
 	"github.com/hypertrace/goagent/sdk/internal/mock"
 	"github.com/stretchr/testify/assert"
@@ -279,15 +279,15 @@ func TestServerRequestFilter(t *testing.T) {
 			body:         "haha",
 			options: &Options{
 				Filter: mock.Filter{
-					URLAndHeadersEvaluator: func(span sdk.Span, url string, headers map[string][]string) filterutils.FilterResult {
+					URLAndHeadersEvaluator: func(span sdk.Span, url string, headers map[string][]string) result.FilterResult {
 						assert.Equal(t, "http://localhost/foo", url)
 						assert.Equal(t, 1, len(headers))
 						assert.Equal(t, []string{"application/json"}, headers["Content-Type"])
-						return filterutils.FilterResult{}
+						return result.FilterResult{}
 					},
-					BodyEvaluator: func(span sdk.Span, body []byte, headers map[string][]string) filterutils.FilterResult {
+					BodyEvaluator: func(span sdk.Span, body []byte, headers map[string][]string) result.FilterResult {
 						assert.Equal(t, []byte("haha"), body)
-						return filterutils.FilterResult{}
+						return result.FilterResult{}
 					},
 				},
 			},
@@ -296,8 +296,8 @@ func TestServerRequestFilter(t *testing.T) {
 			url: "http://localhost/foo",
 			options: &Options{
 				Filter: mock.Filter{
-					URLAndHeadersEvaluator: func(span sdk.Span, url string, headers map[string][]string) filterutils.FilterResult {
-						return filterutils.FilterResult{Block: true, ResponseStatusCode: 403}
+					URLAndHeadersEvaluator: func(span sdk.Span, url string, headers map[string][]string) result.FilterResult {
+						return result.FilterResult{Block: true, ResponseStatusCode: 403}
 					},
 				},
 			},
@@ -307,8 +307,8 @@ func TestServerRequestFilter(t *testing.T) {
 			url: "http://localhost/foo",
 			options: &Options{
 				Filter: mock.Filter{
-					URLAndHeadersEvaluator: func(span sdk.Span, url string, headers map[string][]string) filterutils.FilterResult {
-						return filterutils.FilterResult{Block: true, ResponseStatusCode: 403}
+					URLAndHeadersEvaluator: func(span sdk.Span, url string, headers map[string][]string) result.FilterResult {
+						return result.FilterResult{Block: true, ResponseStatusCode: 403}
 					},
 				},
 			},
@@ -321,8 +321,8 @@ func TestServerRequestFilter(t *testing.T) {
 			body:         "haha",
 			options: &Options{
 				Filter: mock.Filter{
-					BodyEvaluator: func(span sdk.Span, body []byte, headers map[string][]string) filterutils.FilterResult {
-						return filterutils.FilterResult{Block: true, ResponseStatusCode: 403}
+					BodyEvaluator: func(span sdk.Span, body []byte, headers map[string][]string) result.FilterResult {
+						return result.FilterResult{Block: true, ResponseStatusCode: 403}
 					},
 				},
 			},
@@ -364,9 +364,9 @@ func TestProcessingBodyIsTrimmed(t *testing.T) {
 
 	wh, _ := WrapHandler(h, mock.SpanFromContext, &Options{
 		Filter: mock.Filter{
-			BodyEvaluator: func(span sdk.Span, body []byte, headers map[string][]string) filterutils.FilterResult {
+			BodyEvaluator: func(span sdk.Span, body []byte, headers map[string][]string) result.FilterResult {
 				assert.Equal(t, "{", string(body)) // body is truncated
-				return filterutils.FilterResult{Block: true, ResponseStatusCode: 403}
+				return result.FilterResult{Block: true, ResponseStatusCode: 403}
 			},
 		},
 	}, map[string]string{}).(*handler)
